test: Move benchmarks to separate file benchmarks_test.go
